feat: Abstract final calculation method
Previously, clicking the '=' button would calculate the two numbers
that the user entered and display a result based on whatever operator
was stored at the time. The method called still does that, but has been
renamed more verbosely. In addition, it also "soft-resets" the
calculator to allow the user to continue calculating based on the
previous result.

A few functions were created in order to abstract away some of the above
operations.