fix: Prevent the concatenation of results
Prior to this commit, the result of a calculation erroneously absorbed
further inputs by "adding" them to the end instead of processing them
in another calculation. This commit corrects this.

Also redirects the click of an operator button to the calculation step
instead of "blanking" on repeat presses.

Also (this the last of it, I promise) prevents the "equals" button from
being ran before all parts of the calculation are present.