fix: Reorder calculation logic
Previously, post-results calculation worked strangely, treating new
numbers as the first number and the prior result as the second one. This
rewrites some of the logic to only consider the prior result if it is
followed by an operator and then another number. Now, if users would
like to do a brand new operation, they simply start with a number
instead. The clear button is also an option.